+++ packages/rpc_dart_generator/example/bin/main_codec.dart
@@ -2,8 +2,7 @@ import 'package:rpc_dart/rpc_dart.dart';
 import 'package:rpc_dart_generator_consumer/calculator_with_codec.dart';
 
 class CalculatorCodecResponder extends CalculatorCodecContractResponder {
-  CalculatorCodecResponder({String? serviceName})
-    : super(serviceNameOverride: serviceName);
+  CalculatorCodecResponder() : super();
 
   @override
   Future<SumResponse> sum(SumRequest request, {RpcContext? context}) async {
@@ -16,16 +15,11 @@ Future<void> main() async {
   final (callerTransport, responderTransport) = RpcInMemoryTransport.pair();
 
   final responderEndpoint = RpcResponderEndpoint(transport: responderTransport);
-  responderEndpoint.registerServiceContract(
-    CalculatorCodecResponder(
-      serviceName: CalculatorCodecContractNames.instance('json'),
-    ),
-  );
+  responderEndpoint.registerServiceContract(CalculatorCodecResponder());
   responderEndpoint.start();
 
   final caller = CalculatorCodecContractCaller(
     RpcCallerEndpoint(transport: callerTransport),
-    serviceNameOverride: CalculatorCodecContractNames.instance('json'),
   );
 
   final res = await caller.sum(SumRequest(values: [4, 5, 6]));